+++ src/components/clusterselection/EditCluster.vue
@@ -3,7 +3,7 @@
          :title="t('cluster_selection.edit_cluster.edit.title')"
          @click.stop="dialog = true" />
 
-  <q-dialog v-model="dialog">
+  <q-dialog v-model="dialog" transition-duration="100">
     <q-card style="min-width: 800px">
       <q-card-section class="flex justify-between">
         <h2 class="text-h6 q-my-none">
@@ -17,7 +17,7 @@
 
       <q-card-section>
         <form @submit.prevent>
-          <cluster-form-fields v-model="cluster" v-model:formValid="formValid" />
+          <cluster-form-fields v-model="cluster" v-model:form-valid="formValid" />
 
           <q-btn :label="t('setup.test_and_connect.form.test_connection')"
                  :disable="!formValid"