+++ src/components/search/SearchDocuments.vue
@@ -17,8 +17,8 @@
                      autofocus
                      :label="t('search.form.query.label')"
                      @keydown.esc="searchStore.q = '*'" />
-            <!-- eslint-disable-next-line vue/no-v-html -->
-            <small v-html="t('search.form.query.messages')" />
+
+            <search-examples />
           </div>
 
           <div class="col">
@@ -27,6 +27,9 @@
 
           <div class="col-auto">
             <q-btn :label="t('search.form.search')" color="primary-dark" type="submit" class="q-mt-sm" />
+            <div v-if="searchResults?.took" class="text-muted font-13 text-center">
+              {{ searchResults.took }}ms
+            </div>
           </div>
         </div>
       </q-form>
@@ -67,6 +70,15 @@
 
     <loader-status v-else :request-state="requestState">
       <search-results-table :results="searchResults" @request="onRequest" @reload="search" />
+      <template #error>
+        <div class="text-center">
+          <q-btn :label="t('search.form.customize_query.reset')"
+                 size="md"
+                 color="positive"
+                 class="q-ma-md"
+                 @click="resetAndLoad" />
+        </div>
+      </template>
     </loader-status>
   </q-card>
 </template>
@@ -79,8 +91,13 @@
   import SearchResultsTable from './SearchResultsTable.vue'
   import { useSearchDocuments } from '../../composables/components/search/SearchDocuments'
   import { useTranslation } from '../../composables/i18n.ts'
+  import SearchExamples from './SearchExamples.vue'
 
   const CodeEditor = defineAsyncComponent(() => import('../shared/CodeEditor.vue'))
+  const resetAndLoad = () => {
+    searchStore.resetSearchQuery()
+    search()
+  }
 
   const t = useTranslation()
   const {