+++ src/components/shared/LoaderStatus.vue
@@ -11,6 +11,10 @@
             </q-btn>
           </template>
         </q-banner>
+
+        <div>
+          <slot name="error" />
+        </div>
       </div>
     </template>
     <template v-else-if="requestState.networkError">
@@ -25,6 +29,7 @@
   <div v-else>
     <q-linear-progress v-if="requestState.loading" indeterminate />
     <slot />
+    <q-linear-progress v-if="requestState.loading" indeterminate />
   </div>
 </template>