+++ src/services/ElasticsearchAdapter.ts
@@ -72,8 +72,16 @@ export default class ElasticsearchAdapter {
     return this.request(`_cat/indices/${query}`, 'GET', params)
   }
 
-  catIndexTemplates () {
+  templates () {
     return this.request('_template', 'GET')
+  }
+
+  indexTemplates () {
+    return this.request('_index_template', 'GET')
+  }
+
+  componentTemplates () {
+    return this.request('_component_template', 'GET')
   }
 
   catShards (params: object, filter?: string) {
@@ -82,40 +90,40 @@ export default class ElasticsearchAdapter {
   }
 
   indexGetAlias ({ index }: { index: string }) {
-    return this.request(`${index}/_alias`, 'GET')
+    return this.request(`${cleanIndexName(index)}/_alias`, 'GET')
   }
 
   indexAddAlias ({ index, alias }: { index: string, alias: string }) {
-    return this.request(`${index}/_alias/${alias}`, 'PUT')
+    return this.request(`${cleanIndexName(index)}/_alias/${alias}`, 'PUT')
   }
 
   indexDeleteAlias ({ index, alias }: { index: string, alias: string }) {
-    return this.request(`${index}/_alias/${alias}`, 'DELETE')
+    return this.request(`${cleanIndexName(index)}/_alias/${alias}`, 'DELETE')
   }
 
   indexCreate ({ index, body }: { index: string, body?: object }) {
-    return this.request(`${index}`, 'PUT', body)
+    return this.request(`${cleanIndexName(index)}`, 'PUT', body)
   }
 
   deleteByQuery ({ index }: { index: string }) {
     const body = { query: { match_all: {} } }
-    return this.request(`${index}/_delete_by_query?refresh=true`, 'POST', body)
+    return this.request(`${cleanIndexName(index)}/_delete_by_query?refresh=true`, 'POST', body)
   }
 
   indexGet (params: Record<string, any>) {
     const index = Array.isArray(params.index) ? params.index.join(',') : params.index
-    return this.request(`${index}`, 'GET')
+    return this.request(`${cleanIndexName(index)}`, 'GET')
   }
 
   indexStats ({ index }: { index: string }) {
-    return this.request(`${index}/_stats`, 'GET')
+    return this.request(`${cleanIndexName(index)}/_stats`, 'GET')
   }
 
   indexDelete ({ indices }: { indices: string[] }) {
     if (indices.length > MAX_INDICES_PER_REQUEST) {
       return this.callInChunks({ method: 'indexDelete', indices })
     } else {
-      return this.request(indices.join(','), 'DELETE')
+      return this.request(cleanIndexName(indices.join(',')), 'DELETE')
     }
   }
 
@@ -168,11 +176,11 @@ export default class ElasticsearchAdapter {
   }
 
   indexExists ({ index }: { index: string }) {
-    return this.request(`${index}`, 'HEAD')
+    return this.request(`${cleanIndexName(index)}`, 'HEAD')
   }
 
   indexPutSettings ({ index, body }: { index: string, body: object }) {
-    return this.request(`${index}/_settings`, 'PUT', body)
+    return this.request(`${cleanIndexName(index)}/_settings`, 'PUT', body)
   }
 
   reindex ({ source, dest }: { source: string, dest: string }) {
@@ -190,25 +198,35 @@ export default class ElasticsearchAdapter {
     return this.request('_nodes', 'GET')
   }
 
-  index ({ index, type, id, params }: { index: string, type: string, id: any, params: any }) {
-    return this.request(`${index}/${type}/${encodeURIComponent(id)}?refresh=true`, 'PUT', params)
+  index ({ index, type, id, routing, params }: {
+    index: string,
+    type: string,
+    id: any,
+    routing: string,
+    params: any
+  }) {
+    let path = `${cleanIndexName(index)}/${type}/${encodeURIComponent(id)}?refresh=true`
+    if (routing) path += `&routing=${routing}`
+    return this.request(path, 'PUT', params)
   }
 
-  get ({ index, type, id, routing }: { index: string, type: string, id: any, routing: any }) {
+  get ({ index, type, id, routing }: { index: string, type: string, id: any, routing?: string }) {
     const params: IndexGetArgs = {}
     if (routing) params.routing = routing
-    return this.request(`${index}/${type}/${encodeURIComponent(id)}`, 'GET', params)
+    return this.request(`${cleanIndexName(index)}/${type}/${encodeURIComponent(id)}`, 'GET', params)
   }
 
-  delete ({ index, type, id }: { index: string, type: string, id: any }) {
-    return this.request(`${index}/${type}/${encodeURIComponent(id)}?refresh=true`, 'DELETE')
+  delete ({ index, type, id, routing }: { index: string, type: string, id: any, routing?: string }) {
+    let path = `${cleanIndexName(index)}/${type}/${encodeURIComponent(id)}?refresh=true`
+    if (routing) path += `&routing=${routing}`
+    return this.request(path, 'DELETE')
   }
 
   search (params: object, searchIndex?: string | string[]) {
     const index = Array.isArray(searchIndex) ? searchIndex.join(',') : searchIndex
 
     if (index && index.length > 0) {
-      return this.request(`${index}/_search`, 'POST', params)
+      return this.request(`${cleanIndexName(index)}/_search`, 'POST', params)
     } else {
       return this.request('_search', 'POST', params)
     }
@@ -312,3 +330,5 @@ export default class ElasticsearchAdapter {
     }
   }
 }
+
+const cleanIndexName = (index: string) => (index.replace(/%/g, '%25'))