+++ src/store/codeEditor.ts
@@ -1,11 +1,13 @@
 import { defineStore } from 'pinia'
 
 type CodeEditorState = {
-  wrapLines: boolean
+  wrapLines: boolean,
+  vimMode: boolean
 }
 
 export const useCodeEditorStore = defineStore('codeEditor', {
   state: (): CodeEditorState => ({
+    vimMode: false,
     wrapLines: false
   }),
   persist: true