+++ tests/e2e/mocks/elastic_1/indices.ts
@@ -3,7 +3,7 @@ import { Page } from '@playwright/test'
 export const mockElasticIndices = async (page: Page) => {
   const mocks = {
     catIndices: {
-      url: 'http://localhost:9200/_cat/indices/?h=index%2Chealth%2Cstatus%2Cuuid%2Cpri%2Crep%2Cdocs.count%2Cstore.size%2Csc&bytes=b',
+      url: 'http://localhost:9200/_cat/indices/?h=index%2Chealth%2Cstatus%2Cuuid%2Cpri%2Crep%2Cdocs.count%2Cstore.size%2Csc%2Ccd&bytes=b',
       json: [
         {
           'index': 'movies',
@@ -28,8 +28,8 @@ export const mockElasticIndices = async (page: Page) => {
   }
 
   for (const method in mocks) {
-    const url = mocks[method].url
-    const json = mocks[method].json
+    const url = mocks[method as keyof typeof mocks].url
+    const json = mocks[method as keyof typeof mocks].json
 
     await page.route(url, async route => (await route.fulfill({ json })))
   }